--- .env-sample
@@ -1,16 +1,0 @@
-PORT=
-
-DB_URL=
-DB_MAX_OPEN_CONN=
-DB_MAX_OPEN_IDLE=
-DB_MAX_IDLE_TIME=
-
-JWT_SECRET=
-JWT_ISS=
-JWT_AUD=
-
-SMPT_HOST=
-SMPT_PORT=
-SMPT_USERNAME=
-SMPT_PASSWORD=
-SMPT_SENDER=
