+++ README.md
@@ -1,9 +1,5 @@
 # syncwatch-be
 
 Prerequisites:
-Have docker installed to execute tasks. For I'm using https://taskfile.dev/installation
-
-Create DB like so:
-`docker run -d --name db -p 5432:5432 -e POSTGRES_PASSWORD=pswd123 -e POSTGRES_USER=postgres -e POSTGRES_DB=syncwatch postgres`
-
-// fill environment variables
+Have docker installed to execute tasks. For projects task execution I'm using https://taskfile.dev/installation
+instead of make
