+++ Taskfile.yml
@@ -1,13 +1,11 @@
 version: '3'
 
-dotenv: [ '.env' ]
-
 vars:
   GO_VERSION: 1.22
 
 tasks:
   run:
-    desc: 'Run the cmd/api application'
+    desc: run the cmd/api application
     cmds:
       - go run ./cmd/api
 
@@ -29,8 +27,7 @@ tasks:
       - echo 'Linting code...'
       - |
         docker run --rm \
-                   --name linter \
-                   --volume "./:/app" \
+                   --volume "{{.USER_WORKING_DIR}}:/app" \
                    -w /app \
                    golangci/golangci-lint golangci-lint run -v --go {{.GO_VERSION}}
 
@@ -44,8 +41,7 @@ tasks:
       - go fmt ./...
       - |
         docker run --rm \
-                   --name linter \
-                   --volume "./:/app" \
+                   --volume "{{.USER_WORKING_DIR}}:/app" \
                    -w /app \
                    golangci/golangci-lint golangci-lint run -v --fix
 
@@ -56,23 +52,22 @@ tasks:
       - docker image prune --filter label=stage=gobuilder -f
 
   db:create-migration:
-    desc: 'Create a new database migration, usage: task db:create-migration NAME=migration_name'
+    desc: 'create a new database migration, usage: task db:create-migration NAME=migration_name'
     requires:
       vars: [ NAME ]
     cmds:
       - echo 'Creating migration files for {{.NAME}}...'
       - |
         docker run --rm \
-                   --volume "./migrations:/migrations" \
+                   --volume "{{.USER_WORKING_DIR}}/migrations:/migrations" \
                    migrate/migrate create -ext=sql -dir=migrations {{.NAME}}
 
   db:run-migration:
-    desc: 'Run database migrations'
+    desc: 'run database migrations'
     cmds:
       - echo "Running up migrations..."
       - |
         docker run --rm \
-                   --volume "./migrations:/migrations" \
-                   --name db-migrations \
+                   --volume "{{.USER_WORKING_DIR}}/migrations:/migrations" \
                    --network host \
-                   migrate/migrate -path=/migrations -database $DB_URL --verbose up
+                   migrate/migrate -path=/migrations/ -database ${GREENLIGHT_DB_DSN}?sslmode=disable up
