+++ cmd/api/main.go
@@ -1,44 +1,26 @@
 package main
 
 import (
-	"context"
 	"log/slog"
-	"os"
+	"sync"
 
-	"github.com/kiennyo/syncwatch-be/internal/config"
-	"github.com/kiennyo/syncwatch-be/internal/db"
-	"github.com/kiennyo/syncwatch-be/internal/domain/users"
-	"github.com/kiennyo/syncwatch-be/internal/http"
-	"github.com/kiennyo/syncwatch-be/internal/mail"
-	"github.com/kiennyo/syncwatch-be/internal/security"
+	"github.com/kiennyo/syncwatch-be/internal/domain/auth"
+	"github.com/kiennyo/syncwatch-be/internal/infrastructure/config"
+	"github.com/kiennyo/syncwatch-be/internal/infrastructure/http"
+	"github.com/kiennyo/syncwatch-be/internal/infrastructure/log"
 )
 
+var wg sync.WaitGroup
+
 func main() {
-	ctx := context.Background()
+	log.Init()
 
-	cfg := config.Load()
+	server := http.New(&wg, config.HTTP{
+		Port: 3000,
+	}).
+		AddRoutes("/auth", auth.Handlers())
 
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	slog.SetDefault(logger)
-
-	postgres, err := db.New(ctx, cfg.DB)
-	if err != nil {
-		slog.Error("Failed to connect to db", "reason", err.Error()) // Fatal
-		return
-	}
-
-	mailer := mail.New(cfg.SMTP)
-	tokens := security.NewTokenFactory(cfg.Security)
-
-	// users module setup
-	userRepo := users.NewRepository(postgres)
-	userService := users.NewService(userRepo, tokens, mailer)
-	usersHandler := users.NewHandler(userService)
-
-	server := http.New(cfg.HTTP, tokens).
-		AddRoutes("/users", usersHandler.Handlers())
-
-	if err = server.Serve(); err != nil {
+	if err := server.Serve(); err != nil {
 		slog.Error("Failed to start server", "reason", err.Error()) // Fatal
 	}
 }
