+++ internal/domain/auth/handler.go
@@ -1,0 +1,10 @@
+package auth
+
+import (
+	"github.com/go-chi/chi/v5"
+)
+
+func Handlers() chi.Router {
+	r := chi.NewRouter()
+	return r
+}
