+++ internal/domain/auth/service.go
@@ -1,0 +1,1 @@
+package auth
