--- internal/domain/users/entity.go
@@ -1,37 +1,0 @@
-package users
-
-import (
-	"time"
-
-	"github.com/google/uuid"
-	"golang.org/x/crypto/bcrypt"
-)
-
-type user struct {
-	ID        uuid.UUID `json:"id"`
-	Name      string    `json:"name"`
-	Email     string    `json:"email"`
-	Password  password  `json:"-"`
-	Activated bool      `json:"activated"`
-	Scopes    []string  `json:"scopes"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-}
-
-type password struct {
-	plaintext *string
-	hash      []byte
-}
-
-func (p *password) set(plaintextPassword string) error {
-	cost := 12
-	hash, err := bcrypt.GenerateFromPassword([]byte(plaintextPassword), cost)
-	if err != nil {
-		return err
-	}
-
-	p.plaintext = &plaintextPassword
-	p.hash = hash
-
-	return nil
-}
