--- internal/domain/users/errors.go
@@ -1,5 +1,0 @@
-package users
-
-import "errors"
-
-var errDuplicateEmail = errors.New("duplicate email")
