--- internal/http/json/json.go
@@ -1,109 +1,0 @@
-package json
-
-import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"log/slog"
-	"net/http"
-	"strings"
-)
-
-const maxPayloadSize = 1048576
-
-type Envelope map[string]any
-
-type MalformedRequest struct {
-	Status int
-	Msg    string
-}
-
-func (mr *MalformedRequest) Error() string {
-	return mr.Msg
-}
-
-func WriteJSON(w http.ResponseWriter, status int, data any, headers http.Header) error {
-	js, err := json.Marshal(data)
-	if err != nil {
-		return err
-	}
-
-	for key, value := range headers {
-		w.Header()[key] = value
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-
-	_, err = w.Write(js)
-	if err != nil {
-		slog.Error("Write failed", "reason", err)
-	}
-
-	return nil
-}
-
-//nolint:revive,cyclomatic
-func ReadJSON(w http.ResponseWriter, r *http.Request, dst any) error {
-	ct := r.Header.Get("Content-Type")
-	if ct != "" {
-		mediaType := strings.ToLower(strings.TrimSpace(strings.Split(ct, ";")[0]))
-		if mediaType != "application/json" {
-			msg := "Content-Type header is not application/json"
-			return &MalformedRequest{Status: http.StatusUnsupportedMediaType, Msg: msg}
-		}
-	}
-
-	r.Body = http.MaxBytesReader(w, r.Body, maxPayloadSize)
-	dec := json.NewDecoder(r.Body)
-	dec.DisallowUnknownFields()
-
-	err := dec.Decode(&dst)
-	if err != nil {
-		var syntaxError *json.SyntaxError
-		var unmarshalTypeError *json.UnmarshalTypeError
-
-		switch {
-		case errors.As(err, &syntaxError):
-			msg := fmt.Sprintf("Request body contains badly-formed JSON (at position %d)", syntaxError.Offset)
-			return &MalformedRequest{Status: http.StatusBadRequest, Msg: msg}
-
-		case errors.Is(err, io.ErrUnexpectedEOF):
-			msg := "Request body contains badly-formed JSON"
-			return &MalformedRequest{Status: http.StatusBadRequest, Msg: msg}
-
-		case errors.As(err, &unmarshalTypeError):
-			msg := fmt.Sprintf(
-				"Request body contains an invalid value for the %q field (at position %d)",
-				unmarshalTypeError.Field,
-				unmarshalTypeError.Offset,
-			)
-			return &MalformedRequest{Status: http.StatusBadRequest, Msg: msg}
-
-		case strings.HasPrefix(err.Error(), "json: unknown field "):
-			fieldName := strings.TrimPrefix(err.Error(), "json: unknown field ")
-			msg := fmt.Sprintf("Request body contains unknown field %s", fieldName)
-			return &MalformedRequest{Status: http.StatusBadRequest, Msg: msg}
-
-		case errors.Is(err, io.EOF):
-			msg := "Request body must not be empty"
-			return &MalformedRequest{Status: http.StatusBadRequest, Msg: msg}
-
-		case err.Error() == "http: request body too large":
-			msg := "Request body must not be larger than 1MB"
-			return &MalformedRequest{Status: http.StatusRequestEntityTooLarge, Msg: msg}
-
-		default:
-			return err
-		}
-	}
-
-	err = dec.Decode(&struct{}{})
-	if !errors.Is(err, io.EOF) {
-		msg := "Request body must only contain a single JSON object"
-		return &MalformedRequest{Status: http.StatusBadRequest, Msg: msg}
-	}
-
-	return nil
-}
