+++ internal/infrastructure/log/logger.go
@@ -1,0 +1,12 @@
+package log
+
+import (
+	"log/slog"
+	"os"
+)
+
+func Init() {
+	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
+
+	slog.SetDefault(logger)
+}
