--- internal/mail/mailer.go
@@ -1,79 +1,0 @@
-package mail
-
-import (
-	"bytes"
-	"embed"
-	"html/template"
-	"time"
-
-	"github.com/go-mail/mail/v2"
-
-	"github.com/kiennyo/syncwatch-be/internal/config"
-)
-
-type Sender interface {
-	Send(recipient, templateFile string, data any) error
-}
-
-var _ Sender = (*mailer)(nil)
-
-//go:embed "templates"
-var templateFS embed.FS
-
-type mailer struct {
-	dialer *mail.Dialer
-	sender string
-}
-
-func (m *mailer) Send(recipient, templateFile string, data any) error {
-	tmpl, err := template.New("email").ParseFS(templateFS, "templates/"+templateFile)
-	if err != nil {
-		return err
-	}
-
-	subject := new(bytes.Buffer)
-	err = tmpl.ExecuteTemplate(subject, "subject", data)
-	if err != nil {
-		return err
-	}
-
-	plainBody := new(bytes.Buffer)
-	err = tmpl.ExecuteTemplate(plainBody, "plainBody", data)
-	if err != nil {
-		return err
-	}
-
-	htmlBody := new(bytes.Buffer)
-	err = tmpl.ExecuteTemplate(htmlBody, "htmlBody", data)
-	if err != nil {
-		return err
-	}
-
-	msg := mail.NewMessage()
-	msg.SetHeader("To", recipient)
-	msg.SetHeader("From", m.sender)
-	msg.SetHeader("Subject", subject.String())
-	msg.SetBody("text/plain", plainBody.String())
-	msg.AddAlternative("text/html", htmlBody.String())
-
-	for i := 1; i <= 3; i++ {
-		err = m.dialer.DialAndSend(msg)
-		if nil == err {
-			return nil
-		}
-
-		time.Sleep(500 * time.Millisecond)
-	}
-
-	return err
-}
-
-func New(cfg config.SMPT) Sender {
-	dialer := mail.NewDialer(cfg.Host, cfg.Port, cfg.Username, cfg.Password)
-	dialer.Timeout = 5 * time.Second
-
-	return &mailer{
-		dialer: dialer,
-		sender: cfg.Sender,
-	}
-}
