--- internal/mail/templates/user_welcome.gohtml
@@ -1,37 +1,0 @@
-{{define "subject"}}Welcome to Syncwatch!{{end}}
-
-{{define "plainBody"}}
-Hi,
-
-Thanks for signing up for a Syncwatch account. We're excited to have you on board!
-
-{"token": "{{.activationToken}}"}
-
-Please note that this is a one-time use token and it will expire in 3 days.
-
-Thanks,
-
-The Syncwatch Team
-{{end}}
-{{define "htmlBody"}}
-<!doctype html>
-<html>
-    <head>
-        <meta name="viewport" content="width=device-width" />
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-    </head>
-
-    <body>
-        <p>Hi,</p>
-        <p>Thanks for signing up for a Syncwatch account. We're excited to have you on board!</p>
-        <pre>
-            <code>
-                {"token": "{{.activationToken}}"}
-            </code>
-        </pre>
-        <p>Please note that this is a one-time use token and it will expire in 3 days.</p>
-        <p>Thanks,</p>
-        <p>The Syncwatch Team</p>
-    </body>
-</html>
-{{end}}
